fix(dashboard): gate security-schedule fixtures behind DEV mode
Audit finding P12: fabricated bond cash-flow schedule rows (with
Posted/Variance/Forecast posting statuses) were compiled into the
production accounting view-model and rendered as real accounting data
whenever a security had no live trust snapshot — bypassing the
import.meta.env.DEV fixture lane and the operator "Demo data" notice.

- Move the fixture map into lib/security-schedule-dev-fixtures.ts.
- resolveSecurityScheduleEvents returns an empty schedule outside DEV
  and marks development-fixture usage when a fixture is served, so the
  app-shell demo-data banner fires like every other fixture lane.
- Export markDevelopmentFixtureUsage from lib/api.

Verified: vitest accounting-screen.view-model.test.ts (49 passed,
incl. new production-branch and banner-flag tests); tsc strict
typecheck clean.
