+++ src/Meridian.Ui/dashboard/src/screens/accounting-screen.view-model.ts
@@ -31,6 +31,8 @@ import {
 } from "@/lib/workspace";
 import { EXPORT_API_ENDPOINTS, type ReferenceDataWorkbenchEndpointSeed } from "@/lib/workstation-endpoints";
 import { formatReportPackRecipientList } from "@/lib/reporting-distributions";
+import { markDevelopmentFixtureUsage } from "@/lib/api";
+import { resolveDevSecurityScheduleEvents } from "@/lib/security-schedule-dev-fixtures";
 import { formatBytes, formatCount, formatCurrency, formatDateTimeLabel, formatSignedCurrency, toDomId } from "./accounting-screen.formatting";
 import {
   buildSecurityConflictAction, buildSecurityIdentityAliasRow, buildSecurityIdentityIdentifierRow,
@@ -2958,113 +2960,6 @@ const defaultSecurityMasterDrillInServices: SecurityMasterDrillInServices = {
   getInstrumentPassport: (securityId) => getSecurityInstrumentPassport(securityId),
   getTradingParameters: (securityId) => getTradingParameters(securityId),
   getTrustSnapshot: (securityId) => getSecurityTrustSnapshot(securityId)
-};
-
-const securityScheduleFixtures: Record<string, SecurityCashFlowScheduleEvent[]> = {
-  "sec-dev-004": [
-    {
-      eventId: "sched-sec-dev-004-cpn-2026-06",
-      securityId: "sec-dev-004",
-      scheduleFamily: "bond",
-      eventType: "Coupon",
-      paymentDate: "2026-06-15",
-      accrualStartDate: "2025-12-15",
-      accrualEndDate: "2026-06-15",
-      couponRatePct: 5.875,
-      expectedAmount: 29375,
-      actualAmount: null,
-      principalAmount: null,
-      interestAmount: 29375,
-      factorStart: 1,
-      factorEnd: 1,
-      currency: "USD",
-      postingStatus: "Forecast",
-      auditReference: "fixture/security-master/cash-flow/sec-dev-004/cpn-2026-06",
-      note: "Semi-annual fixed coupon projected from the reference coupon schedule."
-    },
-    {
-      eventId: "sched-sec-dev-004-paydown-2026-09",
-      securityId: "sec-dev-004",
-      scheduleFamily: "structured",
-      eventType: "Paydown",
-      paymentDate: "2026-09-15",
-      accrualStartDate: "2026-06-15",
-      accrualEndDate: "2026-09-15",
-      couponRatePct: 5.875,
-      expectedAmount: 148750,
-      actualAmount: 147920,
-      principalAmount: 125000,
-      interestAmount: 23750,
-      factorStart: 1,
-      factorEnd: 0.875,
-      currency: "USD",
-      postingStatus: "Variance",
-      auditReference: "fixture/security-master/cash-flow/sec-dev-004/paydown-2026-09",
-      note: "Principal paydown carries a small expected-versus-actual variance for operator review."
-    },
-    {
-      eventId: "sched-sec-dev-004-maturity-2031-12",
-      securityId: "sec-dev-004",
-      scheduleFamily: "bond",
-      eventType: "Maturity",
-      paymentDate: "2031-12-15",
-      accrualStartDate: "2031-06-15",
-      accrualEndDate: "2031-12-15",
-      couponRatePct: 5.875,
-      expectedAmount: 529375,
-      actualAmount: null,
-      principalAmount: 500000,
-      interestAmount: 29375,
-      factorStart: 0.875,
-      factorEnd: 0,
-      currency: "USD",
-      postingStatus: "Pending",
-      auditReference: "fixture/security-master/cash-flow/sec-dev-004/maturity-2031-12",
-      note: "Final coupon and principal repayment remain pending until trustee schedule confirmation."
-    }
-  ],
-  "sec-1": [
-    {
-      eventId: "sched-sec-1-cpn-2026-05",
-      securityId: "sec-1",
-      scheduleFamily: "bond",
-      eventType: "Coupon",
-      paymentDate: "2026-05-15",
-      accrualStartDate: "2025-11-15",
-      accrualEndDate: "2026-05-15",
-      couponRatePct: 5.25,
-      expectedAmount: 26250,
-      actualAmount: 26250,
-      principalAmount: null,
-      interestAmount: 26250,
-      factorStart: 1,
-      factorEnd: 1,
-      currency: "USD",
-      postingStatus: "Posted",
-      auditReference: "fixture/security-master/cash-flow/sec-1/cpn-2026-05",
-      note: "Validation coupon row used by browser workbench checks."
-    },
-    {
-      eventId: "sched-sec-1-principal-2026-11",
-      securityId: "sec-1",
-      scheduleFamily: "bond",
-      eventType: "Principal",
-      paymentDate: "2026-11-15",
-      accrualStartDate: "2026-05-15",
-      accrualEndDate: "2026-11-15",
-      couponRatePct: 5.25,
-      expectedAmount: 126250,
-      actualAmount: null,
-      principalAmount: 100000,
-      interestAmount: 26250,
-      factorStart: 1,
-      factorEnd: 0.9,
-      currency: "USD",
-      postingStatus: "Pending",
-      auditReference: "fixture/security-master/cash-flow/sec-1/principal-2026-11",
-      note: "Validation amortization row keeps schedule selection consistent."
-    }
-  ]
 };
 
 export function useAccountingCashFlowViewModel(
@@ -6134,11 +6029,19 @@ function buildSecurityStatusAnnouncement({
 }
 
 export function resolveSecurityScheduleEvents(securityId: string | null): SecurityCashFlowScheduleEvent[] {
-  if (!securityId) {
+  // Fixture schedules are a development-only affordance. In production a security without a
+  // live trust snapshot must show an honest empty schedule, never fabricated rows with
+  // posted/variance statuses. The DEV check is read at call time so tests can stub it.
+  if (!securityId || !import.meta.env.DEV) {
     return [];
   }
 
-  return (securityScheduleFixtures[securityId] ?? []).map((event) => ({ ...event }));
+  const events = resolveDevSecurityScheduleEvents(securityId);
+  if (events.length > 0) {
+    // Drives the app-shell "Demo data" notice, matching every other DEV fixture lane.
+    markDevelopmentFixtureUsage();
+  }
+  return events;
 }
 
 export function mapScheduleBookToCashFlowScheduleEvents(
